Enable support for a "default command" https://github.com/oclif/oclif/issues/277
A command can now be designated as the default command (via
package.json -> "oclif": {"defaultCommand": "<cmdId>"}).
When a default command is specified, oclif will attempt to run
that command when it doesn't recognize the command in the input,
or when the input doesn't contain a command, or when
there are no arguments passed to the CLI.

This allows for a command to behave similarly to `yarn [run] <script>`,
where `run` is the default command, such that running `yarn <script>`
does the same thing as `yarn run <script>`.